Fix vote result overlap candidate name
