+++ styles_src/default.scss
@@ -93,23 +93,32 @@ $party-margin: 2px;
     /* candidate including name, party, votes */
     &-text {
       display: flex;
-      justify-content: space-between;
+
+      &--row {
+        flex-direction: row;
+        justify-content: space-between;
+      }
+
+      &--column {
+        flex-direction: column;
+      }
 
       &-candidate {
-        overflow: hidden;
-        white-space: nowrap;
-        text-overflow: ellipsis;
+        width: 100%;
+        display: flex;
+        flex-direction: row;
+        justify-content: space-between;
       }
 
       &-name-container {
+        width: 100%;
         display: flex;
-        overflow: hidden;
-        white-space: nowrap;
-        text-overflow: ellipsis;
       }
 
       &-name {
-        margin-right: 4px;
+        overflow: hidden;
+        white-space: nowrap;
+        text-overflow: ellipsis;
       }
 
       &-party {
@@ -129,7 +138,7 @@ $party-margin: 2px;
       }
 
       &-asterisk {
-        margin-right: 4px;
+        margin-left: 4px;
       }
 
       &-icon-container {
@@ -139,7 +148,14 @@ $party-margin: 2px;
       &-icon {
         height: 0.625em;
         width: 0.625em;
+        margin-left: 4px;
         vertical-align: baseline;
+      }
+
+      &-information {
+        display: flex;
+        flex-direction: row;
+        justify-content: space-between;
       }
     }
